chore: rename MemberBanManager to GuildBanManager
